# [DataTable] Cache boolean result from BinaryNode.EvalBinary