@TrayanZapryanov 2023-11-26
Spoiler: once both PRs are in we should receive following results

| Method      | ConditionsCount | Mean     | Error    | StdDev   | Median   | Min      | Max      | Gen0   | Allocated |
|------------ |---------------- |---------:|---------:|---------:|---------:|---------:|---------:|-------:|----------:|
| SelectByAge | 1               | 200.8 us |  3.89 us |  4.48 us | 199.3 us | 195.4 us | 209.8 us | 4.7506 |  40.12 KB |
| SelectByAge | 5               | 327.1 us |  9.51 us | 10.95 us | 321.6 us | 315.8 us | 353.6 us | 3.8710 |   42.1 KB |
| SelectByAge | 10              | 495.3 us |  8.05 us |  7.13 us | 495.0 us | 484.2 us | 509.1 us | 3.8911 |  44.63 KB |
| SelectByAge | 15              | 723.7 us | 14.60 us | 16.81 us | 716.6 us | 700.1 us | 751.3 us | 2.8818 |  47.05 KB |