@roji 2023-11-29
@TrayanZapryanov it will likely be a while before I have any time to look at this PR. DataTable in general is de-prioritized (and to a certain extent no longer recommended); it's also not particularly suited for high-performance scenarios. So while I'm sure it's possible to find opportunities to (micro-)optimize it, I'm not sure it's worth the time and the churn.

Having said that, the above does show some non-trivial gains and if you feel it's really worth it, I'd be happy to take a look (but again, probably not anytime soon).