@TrayanZapryanov 2023-11-29
@roji Thank you very much for the response. Unfortunately we are using this functionality in our software and everyday millions of operations goes through it. Both PRs looks quite simple and will reduce a lot overhead and make planet more green :). I will wait until you have some time to check, don't worry.