@TrayanZapryanov 2023-11-30
@stephentoub I am sorry to bother you, but if I remember correctly there was a cache of bools in Task. FromResult before. And it was removed soon(this year). Unfortunately I have no more details, but see some similarities with this PR. Is it possible somewhere in .Net ecosystem to be done such static caching for boxing of bools to object? Then my PR will not be needed and will solve the problem in general.