@stephentoub 2023-11-30
> but if I remember correctly there was a cache of bools in Task. FromResult before. And it was removed soon(this year).

It was not removed. Task.FromResult(bool) will return a cached task. That was added in NET 6 (previously it was used only for async methods).