@MichalPetryka 2023-11-30
> static caching for boxing

This would be a huge breaking change and would require an ECMA modification.