@TrayanZapryanov 2023-11-30
Sure, but is there any real use case where you want to distinguish two boxed true/false values?