@TrayanZapryanov 2023-12-01
@EgorBo thank you for pointing me this interesting discussion. Looks like it will not be changed in the .NET, but do you see any value in current PR to cache bools? In this place boxing scales by number of rows and number of conditions, so easily can reach 50k objects per operation. Another question is it valuable to have analyser for boxing bools(maybe also several other values like 0, DateTime.Min/max.....) which proposes static fields as optimization?