@EgorBo 2023-12-01
> but do you see any value in current PR to cache bools?

The optimization itself is not uncommon, I've seen it in a few internal and public projects, but, presumably, it should be applied only where it indeed saves a lot of work for GC in a realistic scenario, so I'd defer the decision to @roji who definitely knows System.Data better)