@roji 2023-12-01
I definitely see this making sense in scenarios where the cached instances are used internally, and the usages are under control of the application/library (as is the case in e.g. Roslyn). However, this is a BCL API which would be handing out these cached instances to arbitrary consuming user code (please correct me if that's not the case), so at least in theory there's the potential of a breaking change here.

And once again, DataTable/DataSet are definitely not considered high-performance APIs, and are not longer being actively evolved, so I'm not sure this is the right fit here.