@TrayanZapryanov 2023-12-01
@roji I understand your concerns about possible breaking change. In my eyes this is totally internal API and result of it is used only in DataExpression(also internal one), but would be good if you check again. You mention already that it is not advisable to use DataTable API, but what is alternative? I know that a lot of desktop software is having grids, that are bound to DataTables and use for free filtering and indexes of it. Sometimes they store 10 to 100000 rows in it and if you apply a simple select over 1 or 2 columns easily can result to 1 million boxing allocations. Decision is your and I will agree with it, just cannot accept that this API is somehow obsolete and should not be touched or modernized.