@roji 2023-12-04
As a general rule, loading a huge (unbounded?) number of rows and binding them to a UI component is in itself problematic practice. In scenarios where arbitrary large data is displayed via a UI, some sort of pagination/progressive downloading needs to be implemented, limiting the number of rows fetched from the database (and transferred to the client) at for every page.

When things are done this way, then the extra allocation overhead of the DataTable bools really should become negligible.  And conversely, in a real-world scenario, downloading and binding a huge number of rows should already be so problematic for performance, that the extra bool allocations also shouldn't matter much.

I admit I'm talking theoretically at this point, since I haven't had time to dive into the PR and the DataTable parts it touches. If you still feel optimizing this is worthwhile, I'll try take a look - but it will be a while before I'm able to do so.