@SamMonoRT 2024-04-09
Sorry to mention, but DataTable/DataSet is not being actively worked on and we will not be taking changes unless absolutely necessary for maintenance/build purposes only.  Closing this PR.
cc @roji